requirements.txt
Flask
mysqlclient
flask-sqlalchemy
python-dotenv
pymysql
flask-wtf
flask-login
email-validator
pytz
